more renaming and folder reorg
