+++ calicam/calibration/capture_volume/helper_functions/get_stereotriangulated_table.py
@@ -13,7 +13,6 @@ import time
 
 from calicam import __root__
 from calicam.cameras.data_packets import PointPacket, FramePacket, SyncPacket
-from calicam.cameras.camera_array_builder import CameraArrayBuilder
 from calicam.cameras.camera_array import CameraArray
 from calicam.triangulate.triangulator import ArrayTriangulator
 
@@ -25,7 +24,7 @@ from calicam.triangulate.paired_point_builder import (
 
 
 
-def get_stereotriangulated_points(camera_array:CameraArray, point_data_path:Path):
+def get_stereotriangulated_table(camera_array:CameraArray, point_data_path:Path) -> pd.DataFrame:
 
     logger.info(f"Beginning to create stereotriangulated points from data stored at {point_data_path}")
     point_data = pd.read_csv(point_data_path)
@@ -99,7 +98,7 @@ def get_stereotriangulated_points(camera_array:CameraArray, point_data_path:Path
                     for key, value in new_table.items():
                         stereotriangulated_table[key].extend(value)
 
-    logger.info(f"Saving stereotriangulated_points.csv to {point_data_path} for inspection")
+    logger.info(f"Saving stereotriangulated_points.csv to {point_data_path.parent} for inspection")
     stereotriangulated_table = pd.DataFrame(stereotriangulated_table)
     stereotriangulated_table.to_csv(
         Path(point_data_path.parent, "stereotriangulated_points.csv")
